tests/wpt/tests/close-watcher/inside-event-listeners.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>

<body>
<script>
test(t => {
  let events = [];
  let watcher = createRecordingCloseWatcher(t, events);

  watcher.oncancel = () => { watcher.destroy(); }

  watcher.requestClose();
  assert_array_equals(events, ["cancel[cancelable=true]"]);

  watcher.requestClose();
  assert_array_equals(events, ["cancel[cancelable=true]"], "since it was inactive, no more events fired");
}, "destroy() inside oncancel");

test(t => {
  let events = [];
  let watcher = createRecordingCloseWatcher(t, events);

  watcher.oncancel = (e) => { e.preventDefault(); watcher.destroy(); }

  watcher.requestClose();
  assert_array_equals(events, ["cancel[cancelable=true]"]);

  watcher.requestClose();
  assert_array_equals(events, ["cancel[cancelable=true]"], "since it was inactive, no more events fired");
}, "destroy() inside oncancel with preventDefault()");

test(t => {
  let events = [];
  let watcher = createRecordingCloseWatcher(t, events);

  watcher.onclose = () => { watcher.destroy(); }

  watcher.requestClose();
  assert_array_equals(events, ["cancel[cancelable=true]", "close"]);

  watcher.requestClose();
  assert_array_equals(events, ["cancel[cancelable=true]", "close"], "since it was inactive, no more events fired");
}, "destroy() inside onclose");

test(t => {
  let events = [];
  let watcher = createRecordingCloseWatcher(t, events);

  watcher.onclose = (e) => { e.preventDefault(); watcher.destroy(); }

  watcher.requestClose();
  assert_array_equals(events, ["cancel[cancelable=true]", "close"]);

  watcher.requestClose();
  assert_array_equals(events, ["cancel[cancelable=true]", "close"], "since it was inactive, no more events fired");
}, "destroy() inside onclose with preventDefault()");

test(t => {
  let events = [];
  let watcher = createRecordingCloseWatcher(t, events);

  watcher.oncancel = () => { watcher.close(); }

  watcher.requestClose();
  assert_array_equals(events, ["cancel[cancelable=true]", "close"]);

  watcher.requestClose();
  assert_array_equals(events, ["cancel[cancelable=true]", "close"], "since it was inactive, no more events fired");
}, "close() inside oncancel");

test(t => {
  let events = [];
  let watcher = createRecordingCloseWatcher(t, events);

  watcher.oncancel = (e) => { e.preventDefault(); watcher.close(); }

  watcher.requestClose();
  assert_array_equals(events, ["cancel[cancelable=true]", "close"]);

  watcher.requestClose();
  assert_array_equals(events, ["cancel[cancelable=true]", "close"], "since it was inactive, no more events fired");
}, "close() inside oncancel with preventDefault()");

test(t => {
  let events = [];
  let watcher = createRecordingCloseWatcher(t, events);

  watcher.onclose = () => { watcher.close(); }

  watcher.requestClose();
  assert_array_equals(events, ["cancel[cancelable=true]", "close"]);

  watcher.requestClose();
  assert_array_equals(events, ["cancel[cancelable=true]", "close"], "since it was inactive, no more events fired");
}, "close() inside onclose");

test(t => {
  let events = [];
  let watcher = createRecordingCloseWatcher(t, events);

  watcher.onclose = (e) => { e.preventDefault(); watcher.close(); }

  watcher.requestClose();
  assert_array_equals(events, ["cancel[cancelable=true]", "close"]);

  watcher.requestClose();
  assert_array_equals(events, ["cancel[cancelable=true]", "close"], "since it was inactive, no more events fired");
}, "close() inside onclose with preventDefault()");

test(t => {
  let events = [];
  let watcher = createRecordingCloseWatcher(t, events);

  watcher.oncancel = () => { watcher.requestClose(); }

  watcher.requestClose();
  assert_array_equals(events, ["cancel[cancelable=true]", "close"]);

  watcher.requestClose();
  assert_array_equals(events, ["cancel[cancelable=true]", "close"], "since it was inactive, no more events fired");
}, "requestClose() inside oncancel");

test(t => {
  let events = [];
  let watcher = createRecordingCloseWatcher(t, events);

  watcher.oncancel = (e) => { e.preventDefault(); watcher.requestClose(); }

  watcher.requestClose();
  assert_array_equals(events, ["cancel[cancelable=true]"]);

  watcher.requestClose();
  assert_array_equals(events, ["cancel[cancelable=true]", "cancel[cancelable=true]"], "since it was still active, more events fired");
}, "requestClose() inside oncancel with preventDefault()");

test(t => {
  let events = [];
  let watcher = createRecordingCloseWatcher(t, events);

  watcher.onclose = () => { watcher.requestClose(); }

  watcher.requestClose();
  assert_array_equals(events, ["cancel[cancelable=true]", "close"]);

  watcher.requestClose();
  assert_array_equals(events, ["cancel[cancelable=true]", "close"], "since it was inactive, no more events fired");
}, "requestClose() inside onclose");

test(t => {
  let events = [];
  let watcher = createRecordingCloseWatcher(t, events);

  watcher.onclose = (e) => { e.preventDefault(); watcher.requestClose(); }

  watcher.requestClose();
  assert_array_equals(events, ["cancel[cancelable=true]", "close"]);

  watcher.requestClose();
  assert_array_equals(events, ["cancel[cancelable=true]", "close"], "since it was inactive, no more events fired");
}, "requestClose() inside onclose with preventDefault()");
</script>
